Assignment/assignment_3_142502013/pyproject.toml
[project]
name = "mlops-assign-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "toml>=0.10.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
